Fix user profile image bug and goods infomation edit bug

+++ controllers/goodsControllers/modifiedGoods.js
@@ -1,6 +1,6 @@
-const fs = require('fs');
-const path = require('path');
-const { GoodsImage, Goods } = require('../../models');
+const fs = require("fs");
+const path = require("path");
+const { GoodsImage, Goods } = require("../../models");
 
 module.exports = async (req, res, next) => {
   try {
@@ -9,31 +9,36 @@ module.exports = async (req, res, next) => {
     const imageFiles = req.files;
     const findImages = await GoodsImage.findAll({
       where: { goodId: id },
-      attributes: ['imagePath'],
+      attributes: ["imagePath"],
     });
-    for (let i = 0; i < findImages.length; i++) {
-      if (
-        findImages[i].imagePath &&
-        findImages[i].imagePath.includes('uploads')
-      ) {
-        fs.unlink(
-          path.join(__dirname, '../..', findImages[i].imagePath),
-          (err) => {
-            if (err) throw err;
-          },
-        );
+
+    if (req.files) {
+      for (let i = 0; i < findImages.length; i++) {
+        if (
+          findImages[i].imagePath &&
+          findImages[i].imagePath.includes("uploads")
+        ) {
+          fs.unlink(
+            path.join(__dirname, "../..", findImages[i].imagePath),
+            (err) => {
+              if (err) throw err;
+            }
+          );
+        }
       }
+      const imageArray = imageFiles.reduce((acc, img) => {
+        const obj = {};
+        obj.goodId = id;
+        obj.imagePath = img.path;
+        acc.push(obj);
+        return acc;
+      }, []);
+      await GoodsImage.destroy({ where: { goodId: id } });
+      await GoodsImage.bulkCreate(imageArray);
     }
-    const imageArray = imageFiles.reduce((acc, img) => {
-      const obj = {};
-      obj.goodId = id;
-      obj.imagePath = img.path;
-      acc.push(obj);
-      return acc;
-    }, []);
+
     await Goods.update({ title, text }, { where: { id } });
-    await GoodsImage.destroy({ where: { goodId: id }, truncate: true });
-    await GoodsImage.bulkCreate(imageArray);
+
     const updateGoods = await Goods.findOne({ where: { id } });
     res.status(200).json({ redirct_url: `/detail/${updateGoods.id}` });
   } catch (err) {
